drbd: Enforce new current UUID written before writes and check peers before generating a new one
Make sure no write makes it to the backing device before new current-UUID
In case we need to generate a new UUID before the next write we need to
make sure that this update to the meta-data block is persistent before
any write is submitted to the local backing disk.

So far the code ensured that for the first write, but a second one
following the first one was potentially submitted the the backend
immediately.

Improve that by keeping the NEW_CUR_UUID flag set until the meta-data
update is finished, and creating a new flag (WRITING_NEW_CUR_UUID)
to ensure only one context is actually doing it.

AND Check peers just befor creating a new current UUID

When a primary node gets isolated it might see one of its peers
still alive while one peers is already gone. Now check the liveliness
of the peers considerd as alive before creating the new
UUID. In case we loose quorum in that process, do not create
the new current UUID.