build: do not generate *-debian.tar.gz
"make prepare_release" generated two tarballs:
- drbd-utils-23.42.0.tar.gz: without the debian/ dir
- drbd-utils-23.42.0-debian.tar.gz: with the debian/ dir

That was because the debian maintainers want clean tarballs, and we used
both in our old build systems. One or the other depending if we built
for Debian-alikes or RPM based distributions.

Then there came our new build system and I decided I only want that:
- drbd-utils-23.42.0.tar.gz: with the debian/ dir
And that the build system internally handles if it uses the Debian part.

That is generated with "make debrelease".

Stop generating these "-debian" tarballs, nobody uses them.

While at it, and so that it does not get boring, introduce "release". So
we now have:

make release
- drbd-utils-23.42.0.tar.gz: without the debian/ dir
  Which we upload and make the Debian people happy.

make debrelease
- drbd-utils-23.42.0.tar.gz: with the debian/ dir
  Which we use in our build system.

And because people are used to "prepare_release", keep it around, while
"release" and "debrelease" are the more reasonable names.