drbd: fix decoding of the volume number in decode_header()
Commit 81e19e86b65db7cbcaab021d8452ff75d9b2472e introduced -1
as a reserved value to select no specific volume, but decoded
the 16 bit unsigned protocol field for the volume number as
a signed 16 bit value, thereby turning all volume numbers
greater than 32767 into invalid values when converting back to
(signed) int, which is what structures like packet_info use
for the volume number.
Fixed by treating 65535 as a reserved value that is interpreted
as -1, while the values in the range [0, 65534] are treated as
valid volume numbers.

Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>
Signed-off-by: Philipp Reisner <philipp.reisner@linbit.com>