drbd: do not create a new UUID upon write when there is no disk
The recent commit db21a275c912 introduced the idea to check peers
before creating a new current UUID. That is great!
If quorum is not configured we need to check if there is a single copy of
the data available. If not, creating a new current-UUID would only cause
confusion when the lost nodes join again.