drbd: Adjust rs_total if blocks become out of sync during resync
rs_totoal is recorded when the resync begins. The totoal number
of set bits is simply copied over to rs_total. When we add bits
while the resync is already running we should increase rs_total
as well.
That is important since we have a progress percentage now. Without
doing that the progress might turn into a negative number!
On top of that the math in drbdsetup does not expect a negative number
and then display a unreasonable large percentage number.

Closes LINBIT/drbdtop#7