drbd: consider a SyncTarget nodes as stable
A node that is sync source, and then becomes resync target from a newer
peer, pauses the downstream resync.  After it finished its resync from its
upstream it continues the downstream resync.
Now, the previous commit allows us to propagate the D_UP_TO_DATE downstream.

Therefore the this change in definition of the "stable" bit is necessary
and makes sense.