Fix possible race in dtr_select_path_for_tx()
So far the function returned a path, and after it returned we used
dtr_path_get_cm() to get a reference on the cm. That problem was
that the patch could go away in the mean time.
Improve that by returning the cm and calling __dtr_path_get_cm()
within the same rcu_read_lock()