drbd: Improve the resync controller for fast back end devices and network
So far we had the issue that the resync controller was not able to scale to
resync speeds possible with recent storage devices like NVMe drives.

In order to deal with such scenarios more gracefully, reschedule the
rs_make_request() function early in case the resync source fulfilled all
resync requests before the resync timer expired (rs_sect_in reaches the
value of rs_in_flight).

We scale the sect_in value with the "earliness". I.e. if it took only
half of the time, sect_in is multiplied by 2.

But it still consumes a whole slot from the the fifo. With that we narrow
the timing characteristics of the controller a bit, but no problem since
the data flows faster than expected anyways.