Add a timer to detect if a posted tx does not complete in time
It turns out that the timeout mechanism built into rdma/vers is a
jungle. The timeout can only be set if the QP is in SQD state. SQD
state has a different meaning on ib/RoCE than on iWARP. No one on
the kernel tries to use the timeout mechanism of rdma/verbs.
I guess everone simply relies on Linux timers.