Replace the stream_nr with the complete immediate in dtr_tx_desc
This is necessary in order to make the reposting of a dtr_tx_desc
after a transmit timeout possible. We need to repost it with the
same sequence number!

With that the assignment of the sequence number has to be "moved up"
in the call graph, so that __dtr_post_tx_desc() no longer assigns
the sequence number.