drbd: fix cross-resource resync-after dependencies
resync-after dependencies to serialize resync of different volumes
only worked with dependencies between volumes of the same resource.

With cross-resource dependencies,
the resync of the "paused-for-dependency" volume was never resumed,
because we looked at the state of the wrong peer-device
(usually the to-be-resumed peer-device itself).

The minimal fix for this issue is to replace the lookup-by-vnr
in the wrong connection by an iteration over all peer devices
of the device we are interested in.

At some point we may have to dump the global device list walk
and replace this with something more efficient.