drbd: allow disabling usermode helper
This change is intended to help containerization of DRBD components.
Especially the ones that interact with the kernel.

The goal is to even load the kernel module from a container, but for now
assume we loaded the ko on the host. Then, in the container, we create
resources and therefore interact with the host kernel. One thing the
host kernel does is it calls various usermode helpers.

The mechanism is to call the usermode helper (basically
unconditionally), and then userspace decides which handler to run.  But
in a containerized world, we do not want to have "state" on the host
(drbd-utils + handlers).

So in a first step allow disabling the usermode helper. This works
because the module has a "usermode_helper" param.

In a second step we can think about an interface (magic file in /proc,
/sys, whatever) that we can use for host-container communication. Or the
container registers itself (container ID) via a special file, and host
scripts then call into the container via "docker exec" or anything. This
is a second step if we really want handlers in the container.