debian: do not background depmod
This took me quite a long time to figure out why loading the module from
a container did not work.

First it failed on modprobe because depmod was not done.
Then I added a second "depmod -a" (actually postinst does one), which
collided with the first one.

Then I did what every good engineer does, I added sleeps and it worked.

The fix is to not background the depmod.