drbd: Access super block in md directly if possible with DAX/PMEM
In case the meta-data is on persistent memory (PMEM aka NVDIMM) and we
have direct access (DAX) to it, it does not make sense to access it only
by writing/reading blocks from/to it.
We can access it byte-granular, and access can happen from IRQ contexts,
and is nearly as fast as access to main memory.

We just need to decode/encode our "superblock".

Signed-off-by: Philipp Reisner <philipp.reisner@linbit.com>
Co-developed-by: Joel Colledge <joel.colledge@linbit.com>
Signed-off-by: Joel Colledge <joel.colledge@linbit.com>