drbd: Optimize activity_log for DAX/PMEM
In the case of DAX/PMEM we store the activity in it's native representation
as table.

The drbd_al_begin_io_fastpath() function, which is used from the submitter
thread, now does the transaction inline, as long as a AL-entry is available.

No code-path may reach al_write_transaction(), since that would write
an AL-transactions for block-IO media.

First measurements on a 20G NVMe volume with 4K randwrite iodepth=1
Backing disk without DRBD: 9400 IOPS
Metadata on pmem, with DAX: 8200 IOPS
Metadata on pmem, without DAX: 5800 IOPS
Metadata on separate NVMe: 5500 IOPS

This is a massive improvement for write intensive workloads

Signed-off-by: Philipp Reisner <philipp.reisner@linbit.com>
Co-developed-by: Joel Colledge <joel.colledge@linbit.com>
Signed-off-by: Joel Colledge <joel.colledge@linbit.com>