drbd: Optimize access to bitmap in case meta-data is on DAX/PMEM
So we do not allocate DRAM for the bitmap, we do not read it, we never
write it back. Directly access it in place.

Quick comparison measurement with the peer offline shows

Backing disk and metadata on pmem, with DAX, offline peer: 197000 IOPS
Backing disk and metadata on pmem, without DAX, offline peer: 54000 IOPS

Signifficant improvement!

Signed-off-by: Philipp Reisner <philipp.reisner@linbit.com>
Co-developed-by: Joel Colledge <joel.colledge@linbit.com>
Signed-off-by: Joel Colledge <joel.colledge@linbit.com>