drbd: fix potential double call of drbd_backing_device_free
During down/detach, we schedule "DESTROY_DISK" on the DRBD resource worker.
But we also schedule a generic work drbd_device_finalize_work_fn() from the
drbd_destroy_device() destructor.

Both call
	drbd_backing_dev_free(device, device->ldev);
	device->ldev = NULL;

If due to some race both run at "exactly" the same time,
we get a "double free", two calls to bd_unlink_disk_holder(), and
| kernfs: can not remove 'dm-0', no directory
| WARNING: CPU: 0 PID: 4468 at fs/kernfs/dir.c:1504 kernfs_remove_by_name_ns

So we better call that only once.

The backing device life time is unrelated to device life time,
so the correct place is when detaching from the backing device,
not during device destruction.

But we need to wait for the worker to actually process the DESTROY_DISK.

Once (interruptible) from adm_detach(), so a not interrupted
administrative detach will only return after the backing device was
in fact released.

And once in adm_del_minor(), before unregistering the device object,
because otherwise with a quick "detach, del_minor" (aka: "down")
we have a race[*] where the worker might not find the device object
anymore that was scheduled to have its backing device detached.

This wait needs to be uninterruptible.

[*] that race would lead to drbd_backing_device_free() never being called, and
| vgchange -an scratch
|  /sys/block/drbd100/dev: fopen failed: No such file or directory
|  /dev/scratch/s0_00: failed to find associated device structure for holder /dev/drbd100.
|  Logical volume scratch/s0_00 is used by another device.
|  Can't deactivate volume group "scratch" with 1 open logical volume(s)