drbd: quorum: Use a third, diskless node as tiebreaker
The fundamental problem with two node clusters is that in the moment
they lose connectivity we have two partitions and none of the two has
quorum, i.e. the cluster halts the service.

DRBD would block IO (or throw IO errors, depending on configuration)
if the quorum feature is enabled and quorum is lost.

This allows a 3rd, diskless node to act as a quorum tie-breaker.

A partition that has quorum and is about to lose quorum
(either node in a two node cluster, when they lose their
interconnect) will keep quorum as long as it still is
connected to the majority of the diskless nodes in
the system (or the only one).

Let's go through it by example. Node A is a diskful primary, B is a
diskful secondary, and C is the diskless tiebreaker node.

Case 1

    C             C
   / \     ->    / \
  /   \         /   \
 A --- B       A -x- B

Let's say the network between the two diskful nodes is interrupted,
but both can still talk to the diskless, then both keep quorum. If
one of them is primary, the other's disk will be marked as outdated,
i.e. it does not allow service migration. Good!

Case 2

   C             C
  / \     ->    / x
 /   \         /   x
A --- B       A -x- B

The diskless tiebreaker forms a partition with the primary. It
keeps quorum, and the other partition loses quorum since it does
not have the diskless tiebreaker. The secondary's disk will be
UpToDate, but the node will be unable to promote to primary since
it doesn't have quorum.

Case 3

   C             C
  / \     ->    x \
 /   \         x   \
A --- B       A -x- B

The diskless tiebreaker forms a partition with the secondary. The
partition of the primary loses quorum and starts throwing IO errors.
The partition of the secondary keeps quorum, and the service might
get started there.

In addition, if the diskless tries to jump from one partition to the
other we need to avoid data divergence. Let's look at the possible
cases:

Case 1a

   C             C
  / \     ->    x \
 /   \         x   \
A -x- B       A -x- B

The partition of one disk with data plus one diskless runs the
service and suddenly loses the diskless, loses quorum, and stops
the service. Now the diskless node joins forces with the other node,
but nothing can move there since the other node's disk is Outdated.
This results in the entire cluster locking up, so no data divergence
can occur.

Case 2a

   C             C
  / x     ->    x \
 /   x         x   \
A -x- B       A -x- B

The diskless tiebreaker disappears from the working partition and
joins node B (which lost quorum earlier). The result is that node A
loses quorum as well. It is important to note that a node can *not*
regain quorum by adding a diskless node. In other words, the
tiebreaker mechanism only works if the node in question already has
quorum. This, again, locks the whole cluster to ensure data integrity.

Case 3a

   C             C
  x \     ->    / x
 x   \         /   x
A -x- B       A -x- B

Just as above, the tiebreaker disappears from the working partition
and joins the diskful node without quorum (B). Again, A cannot regain
quorum by simply re-adding the diskless tiebreaker, and so the
cluster halts again.