drbd: fix use after potential free (connection breaks while "Ahead")
Once a sender thread stops (or gets restarted), the sender still needs
to cancel sending of requests already queued up on the transfer_log.

connection->todo.req and todo.req_next are caching pointers
to the "currently (or last) processed" and the "next in line" request.

todo.req is assigned from todo.req_next in tl_next_request_for_connection().

todo.req_next is advanced implicitly by the request-state machinery
in __req_mod() and mod_rq_state().

So we better call tl_next_request_for_connection() *after*
__req_mod(req, SEND_CANCELED, ...), not before.

Or we end up calling into __req_mod(req, SEND_CANCELED, ...) twice
for each request.  Which usually is harmless, as long as the request
object stays around, so no-one noticed.
But it can lead to access after free, if the SEND_CANCELED event triggers
recycling of the request object. Which is usually the case if the
connection breaks while we are "AHEAD" and have to cancel send-out-of-sync.

Depending on how fast the freed request object is recycled, this could
lead to endless loops or NULL dereferences or other bad behavior.