drbd: don't use BVEC_ITER compatibility macros
In Linux 3.14 the way to iterate a bio changed. Among other things, the type of
the bio_vec was changed from a regular struct bio_vec to a pointer. This
required the use of several compat macros which can now be removed because
the backwards compatibility is now handled by a coccinelle patch.