drbd: fix invalid iterator usage in rcu contexts
When a list is iterated in an rcu_read_lock context, the respective
iterator macro with the _rcu suffix must be used. This is not the
case in some places, so fix these instances.

This change was made automatically using the following coccinelle script:

// find wrong instances
@ r @
iterator it !~ ".*_rcu$";
@@
rcu_read_lock();
...
it(...) {
	...
}
...
rcu_read_unlock();

// append _rcu
@ script:python py @
it_fixed;
it << r.it;
@@
coccinelle.it_fixed = it + '_rcu'

// replace found instances
@@
iterator r.it;
identifier py.it_fixed;
@@
rcu_read_lock();
...
- it
+ it_fixed
  (...) {
...
}
...
rcu_read_unlock();