build: don't sort coccinelle patches
In order to apply patches "in layers" we need to be able to control
the order in which they are applied. This can be done by manipulating
the order in which the patch names are generated in gen_patch_names.c,
but that requires that we keep the order and don't sort them before
applying.