drbd: remove compat code for hlist_for_each_entry
This compatibility wrapper for hlist_for_each_entry can be
removed, firstly because this is now handled by a coccinelle patch,
and also because it wasn't supposed to be there in the first place.
There were two separate tests that probed for the same behaviour,
so one of them was always redundant.