drbd: first set_in_sync, then send_ack
If we send out a P_RS_WRITE_ACK (or any block ack), the peer may process the
next request, and may chose to send a P_OUT_OF_SYNC.

In theory, that P_OUT_OF_SYNC could be received and processed in the receiver
thread before the ack_sender reaches the "set_in_sync" associated with the
P_RS_WRITE_ACK, and the respective bit could end up being "in sync" in error.

Currently this is an impossible race, because any request sent as P_OUT_OF_SYNC
would be in a different epoch than any request that could cause a
P_RS_WRITE_ACK, which means they are separated by a P_BARRIER,
receive_Barrier() would wait until done_ee_cnt reaches 0, which implies that
the e_end_block() did run in its entirety.

But it feels better to have the calls to change the bit and send the ACK
in the right order anyways.