drbd: introduce P_RS_CANCEL_AHEAD
If we have on-congestion pull-ahead, and a "high" congestion-fill, we may pull
ahead during resync, while the peer is still SyncTarget, and will be for quite
a while, because it will only notice once it receives the state change
notification queued after all the other data that lead to the congestion-fill
and pull-ahead.

To avoid useless back-and-forth of retrying the same resync requests over and
over again, always getting an P_RS_CANCEL in return, and the obviously unwanted
side effect of actually processing these P_RS_DATA_REQUEST by reading the data
from the backend before replying with P_RS_CANCEL anyways, use the new
P_RS_CANCEL_AHEAD instead, which should then immediately cause the peer to stop
sending resync request.