drbd: Fix stuck resync when many resync requests are cancelled
This fixes a bug where resync got stuck indefinitely after a number of
resync requests were cancelled. The problem occurred when
make_resync_request reached the end of the bitmap. The field
rs_in_flight is updated optimistically and then corrected once the
resync requests have been sent. However, it was not corrected on
reaching the end of the bitmap. This caused the resync controller to
incorrectly believe that there were resync sectors in-flight, which
prevented it from requesting any more sectors.

Fix the bug by simplifying the logic so that the correction is always
applied.

Also fix an off-by-one error affecting the number of consecutive blocks
requested.