Makefile/Kbuild: Implement caching of compat patches
Coccinelle/spatch work great for our needs to keep compatibility
with older/distro kernels. Since we can not assume that coccinelle
is available in all distros we want to build nor we want that
users need to install coccinelle in their build environments,
we want to ship pre-computed patches in the release tar.gz.

Introduce the infrastructure for keeping cached cocci results.