drbd: quorum: fix diskless majority calculation
For the quorum tiebreaker feature, we need to figure out whether or not we
can see the majority of the diskless nodes in the cluster. This calculation
previously used the "quorum" config setting to specify the calculation mode
(majority, all, or a fixed number). However, when looking at diskless nodes,
we always want to know if we're seeing the majority of all available nodes,
so always use that mode.