drbd: Fix detection of unstable resync with L_WF_BITMAP_T peer
A primary peer with state L_WF_BITMAP_T cannot be considered a stable
source because it does not send data. The logic to determine whether a
peer is a stable source now matches up with the decision of whether to
send data in drbd_should_do_remote().

This fixes a bug where out-of-sync blocks failed to sync. The sequence
of events was as follows:
* Node A is primary, Node B is a connected secondary, Node C is
  disconnected
* C connects to B, starts bitmap exchange
* C connects to A, starts bitmap exchange
* Sync starts B -> C; due to the bug this is considered a stable sync
* Sync finishes B -> C, disk state would be D_UP_TO_DATE, but is
  corrected to D_OUTDATED by sanitize_state()
* A treats the state change as a loss of contact to peer data (arguably
  another bug), generates a new UUID "X", and sends it to B and C
* Sync runs successfully A -> C
* C disconnects again
* A does not generate a new UUID because the only disconnected peer C
  already has a bitmap UUID
* C re-connects, but is determined to be up-to-date without any sync
  occurring, because it has the UUID "X"