drbd: Scan bitmap per-peer instead of once for the device
This is relevant when we have multiple sync sources from which we sync
in turn. For instance, when we resync from a secondary and have a
primary in paused sync source state. When the unstable resync from the
secondary is finished, we then switch to sync from the primary.

This fixes an issue where resync stalled as follows:
* Unstable resync with a secondary peer is ongoing
* We exchange bitmaps with the primary
* While exchanging bitmaps, P_PEERS_IN_SYNC is received as part of the
  ongoing resync
* When the bitmap exchange is complete, bits in the regions for which
  P_PEERS_IN_SYNC were received are marked dirty
* The unstable resync finishes
* We start to sync from the primary, starting from the device level
  resync position from the first resync, just rounded down
* The sync stalls because it reaches the end of the bitmap without
  syncing the dirty bits from the bitmap exchange