drbd: Use RCU transfer log traversal instead of holding reference to next
This fixes an issue where requests were removed from the transfer log
twice. Transfer log access is no longer synchronized, so a request might
be removed from the transfer log between finding it on the list and
getting a kref to it. Then when the kref is put, it is removed from the
transfer log again.

With RCU this is no longer necessary. We can traverse the list under the
RCU lock without being concerned about the requests being freed.

Warning: Requests that are being re-used for sending peer acks might be
found in such a list traversal.