drbd: quorum: check if intentionally diskless
Consider the following situation:

  A
 /
B---C

In this case, A only sees one of the two other voters. Thus, quorum only
depends on the local disk; if the local disk fails, we will lose quorum.

However, in reality losing the disk means that the disk state goes to
diskless, which in turn means that we count ourselves as a diskless node.
This effectively means that we use ourselves as a quorum tiebreaker, which
is not desirable in this situation.

To fix this, only count ourselves as a diskless node when it was actually
intended in the configuration.