drbd: throw IO errors on all devices
When drbd is configured to throw IO errors upon loss of quorum, and
there are multiple volumes in a resource, the errors will only be thrown
for writes to the volume that lost quorum.

The correct behaviour would be to block all writes to all volumes in the
resource when quorum is lost on any of the volumes.

To implement this efficiently, cache the aggregated result of whether or
not all devices have quorum and use it when deciding whether or not to
throw an IO error.