drbd: fix reconciliation resync vs after-weak resync
Nowadays there are two methods how a resync can be triggered after a primary
node left. If the reconciliation method starts a resync it needs to have
precedence. In that case the after-weak method will not trigger a resync.
This patch just ensures that the no resync decission does not interfer
with the already started resync.