docker: allow building from src + UBI
This keeps the existing logic that one can specify a distribution/node
hash combination (or bind a repo file) to fetch kernel packages from
LINBIT repos.

This also will try to build the kernel module if none of the above
options is given.

On Debian based hosts this should be executed as:
docker run -it --rm \
	-v /lib/modules:/lib/modules:ro \
	-v /usr/src:/usr/src:ro \
	-v /usr/lib:/usr/lib:ro --privileged drbd9:bionic

On RHEL based hosts:
docker run -it --rm \
	-v /lib/modules:/lib/modules:ro \
	-v /usr/src:/usr/src:ro \
	--privileged drbd9:rhel7

Note that you are *not* allowed to bind '/usr/lib'. Binding it is not
necessary and it would only work if you have "gcc" installed on the host
as well.

While at it, add the necessary meta information for OpenShift.