build: differentiate between development and release build
When we are on a developer machine, we want to regenerate the compat patch
(i.e. run coccinelle) only when it is necessary. The problem is that, when
making a release, the sources get packed into a tarball. This can (depending
on the tar implementation) mess up the timestamps and cause the patches to be
inadvertently rebuilt.

Thus, when there is no ".git" directory present, we assume that we came out of
a release tarball and drop the dependency from the patches to the source files.