drbd: Check peers before creating new current UUID
When a primary looses a peer, it tries to resume IOs that where blocked
because we were waiting for acks from the peer we just lost.
Before DRBD may do that, it needs to check if a new UUID needs to be created.
That is allowed only if we still have quorum. Better check in such a situation
if the peers we beliefe are online are really online.