drbd: Delay local promote if primary was lost shortly
If the cluster lost a primary very short ago, it might be still
in process of all the disks going through D_CONSISTENT -> D_UP_TO_DATE
again.

This process also runs concurrently with eventually happening
reconciliation resyncs.

When we start a promotion concurrently it might be that we get
it allowed in phase one but locally rejected at commit time,
because we receive the updates that the peers at D_CONSISTENT
at exactly that time.

It does not hurt to delay the promotion a little bit, after
the above described things happened.