compat test and patch for policy in struct genl_ops
This is the first user of another compat mechanism, which is quite similar
to the Coccinelle patching system, but it bypasses Coccinelle. In some cases
it is simply more desirable to have a plain patch rather than a Coccinelle
patch, such as when Coccinelle is unable to parse the affected code.