build: support for regular patches for compat
Coccinelle patches for kernel compatibility are great, but under some
circumstances a regular patch is more appropriate. One example is when
coccinelle is simply unable to parse the affected code, because it contains
too many macro constructs, etc.

Also, patches are now generated relative to the base directory, not the source
directory.