drbd: Abort two-phase-commits that can not see a commit packet
In case the last connection is lost a prepared transaction can not see a commit
nor an abort packet, therefore cancel the transaction.

If it is a directed transaction and the connection to the initiator of
the transaction is lost also cancel the transaction. The reason for that
is that the only directed transactions are used to establish a connection!

This issue became apparent when you tried to connect two primaries
and dual-primary is not allowed. Then the connect is aborted and the
target node of the 2PC (for establishing the connection) is left with
a prepared transaction.
All following 2PCs get rejected for 2 minutes! (2PC timeout)

With this change following 2PCs can proceed immediately