build: fix compat patching error for rpm
When we generate the compat patches, we assume the directory structure to
look like this:

drbd-version/
    drbd-headers/
        drbd_meta_data.h
	drbd_protocol.h
	...
    drbd/
        drbd_actlog.c
        drbd_bitmap.c
        drbd_dax_pmem.c
        ...

That is to say, we assume the directory containing the source files to be
called "drbd".
The problem is that the rpm build copies out the sources to an "obj"
directory, and renames the directory containing the sources in the process.
To work around this, create a symlink to the source directory that is
called "drbd", which should make patching work again.