drbd: document intentional fall-throughs
With Linux 5.3, the -Wimplicit-fallthrough gcc option is globally enabled
in the kernel Makefile. This change makes a few warnings pop up when
compiling drbd.
Most of the fall-throughs in the code were actually marked with comments
already, but not in a format that gcc recognizes.
So, this change either adds "Fall through" comments where they are necessary,
or formats them so that gcc silences the warning.