compat: remove compat code for blk_queue_merge_bvec
The whole concept of blk_queue_merge_bvec was removed in upstream at
some point, so now our codebase is cluttered with "mostly dead" code
and ifdefs to go with it.

To handle this more nicely, a compat patch is introduced that adds the
code in question when it is needed.