compat: remove compat for kmap_atomic with 2 arguments
k[un]map_atomic() changed from explicitly passing a km_type to determining
it implicitly. Since removing this argument loses semantic information, we
have come up with a mechanism to preserve this information through comments.
See the coccinelle patch in the associated drbd-kernel-compat commit
for an explanation of how it works.