compat: remove compat for alloc_ordered_workqueue
At some point in the history of the kernel, alloc_[ordered_]workqueue evolved
to accept printf-style fmt and args arguments. Previously, the presence of
this functionality was probed by just testing if the Linux version was >3.3.

To solve this issue and still keep the code readable, introduce a compat test
and a coccinelle patch to transform the relevant calls.