drbd: do not set the device RO
So far we did set_disk_ro() depending on the DRBD state (e.g., when
Secondary).

We basically assumed that programs that want to use a DRBD block device
will open() it. This then auto-promotes the DRBD device to Primary, it
is RW, and everything is fine.

Recent versions of kvm (userspace) use some more advanced detection
logic to detect if a device is RO/RW to cover users that set their
devices to RO via blockdev(8). This basically boils down to:
 ioctl(fd, BLKROGET, &readonly)

So we as DRBD never saw an open(), we don't promote, kvm sees us RO and
fails.

Get rid of that whole thing. We consider setting the device RO via
blockdev(8) a user - and user only - decision. It was just kept in the
code base for this long for historic reasons.

Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>