drbd: Align new-current-UUID behaviour between diskless and diskful nodes
When a node with disk becomes primary it only creates a new current-UUID if
it has an empty bitmap slot towards one if its peers.

A diskless node that toggles between secondary and primary multiple
times creates a new current-UUID every time it becomes primary.

That can lead to a false-positive split brain detection if
a 3rd node is absent during that time.

Bring this in sync with the code when a node with disk becomes primary.