drbd: Fix wrong state change when UUID flags change during UUID exchange
This fixes a similar issue to 7d15e1896e93dad4fad35b9e68b5ae3390e31983.

When DRBD exchanges UUIDs after establishing a conneciton. It sends some
flags and expects to receive those from the peer very soon. When it gets
them it compares the peer's UUIDs and flags with its own.

The problem is if the flag values change between when we send them and
when we get those from the peer. They can change because we are doing
a resync on another connection.

When that happens you see one end of a connection in
replication:WFBitMapS and the other end in replication:Established, or
similar mismatched states.

Fix this issue by caching the flags we presented to the peer and
validate them when we compare the UUIDs and flags with those from the
peer.