drbd: Store a value of -1 in unused bitmap slots
In ancient times it was okay to to a bitmap value of 0 to indicate unused.
But since quite some time a value of 0 is used to detect that two machines
are in sync.
For the logic in update_bitmap_slot_of_peer() it is right to use -1 for
unused bitmap UUIDs, also this change prepares the next patch.