drbd: Never keep a bitmap uuid value if there is no bitmap for that peer
Commit 0af5803b13 "Do not have a bitmap_uuid for intentional diskless nodes"
introduced the idea to keep bitmap slots empty that are associated with
intentional diskless nodes (i.e. we do not have a change tracking bitmap
for them).

With commit 4495140002 "fix spurious new current UUID generation" this
detail was lost.

Restoring the idea to never keep an bitmap uuid value for a intentional
diskless node.