drbd: Consider different current-uuids in receive_peer_dagtag()
In most cases the UUIDs are equal, but it might be that one of the
secondaries got one additional new-current-UUID from the primary
before he also lost contact to the primary.
If that is the case we will find the usual pattern that the
other secondaries current-UUID is in my own bitmap slot for
that peer.